Revision: moth--main--1.0--patch-3 Archive: stig@brautaset.org--arwen-03 Creator: Stig Brautaset <stig@brautaset.org> Date: Mon Feb 17 10:39:22 GMT 2003 Standard-date: 2003-02-17 10:39:22 GMT Summary: ability to detect the end of a game and report the winner Keywords: New-files: {arch}/moth/moth--main/moth--main--1.0/stig@brautaset.org--arwen-03/patch-log/patch-3 Modified-files: moth.c New-patches: stig@brautaset.org--arwen-03/moth--main--1.0--patch-3
Added ability to detect the end of a game and report the winner.
We do this simply by counting the number of pieces by each player.

Also merged the two functions used for evaluation of a state. This
operation however included splitting out the heuristic variable -- this
is now a const global variable.